Resolve "Refactor managed datasets to a managed schema"

Closes #174

See merge request MAGIC/ops-data-store!100

+++ resources/data/dataset-schemas.sql
@@ -1,6 +1,12 @@
+-- [SCHEMA]
+
+CREATE SCHEMA IF NOT EXISTS magic_managed;
+
+SET search_path TO magic_managed, public;
+
 -- DEPOT
 
-CREATE TABLE IF NOT EXISTS public.depot
+CREATE TABLE IF NOT EXISTS magic_managed.depot
 (
   pk                         INTEGER GENERATED ALWAYS AS IDENTITY
     CONSTRAINT depot_pk PRIMARY KEY,
@@ -43,23 +49,23 @@ CREATE TABLE IF NOT EXISTS public.depot
 );
 
 CREATE INDEX IF NOT EXISTS depot_geom_idx
-  on public.depot using gist (geom);
+  on magic_managed.depot using gist (geom);
 
 CREATE OR REPLACE TRIGGER depot_updated_at_trigger
   BEFORE INSERT OR UPDATE
-  ON depot
+  ON magic_managed.depot
   FOR EACH ROW
   EXECUTE FUNCTION set_updated_at();
 
 CREATE OR REPLACE TRIGGER depot_updated_by_trigger
   BEFORE INSERT OR UPDATE
-  ON depot
+  ON magic_managed.depot
   FOR EACH ROW
   EXECUTE FUNCTION set_updated_by();
 
 -- INSTRUMENT
 
-CREATE TABLE IF NOT EXISTS public.instrument
+CREATE TABLE IF NOT EXISTS magic_managed.instrument
 (
   pk                         INTEGER                  GENERATED ALWAYS AS IDENTITY
     CONSTRAINT instrument_pk PRIMARY KEY,
@@ -99,23 +105,23 @@ CREATE TABLE IF NOT EXISTS public.instrument
 );
 
 CREATE INDEX IF NOT EXISTS instrument_geom_idx
-  on public.instrument using gist (geom);
+  on magic_managed.instrument using gist (geom);
 
 CREATE OR REPLACE TRIGGER instrument_updated_at_trigger
   BEFORE INSERT OR UPDATE
-  ON instrument
+  ON magic_managed.instrument
   FOR EACH ROW
   EXECUTE FUNCTION set_updated_at();
 
 CREATE OR REPLACE TRIGGER instrument_updated_by_trigger
   BEFORE INSERT OR UPDATE
-  ON instrument
+  ON magic_managed.instrument
   FOR EACH ROW
   EXECUTE FUNCTION set_updated_by();
 
 -- WAYPOINT
 
-CREATE TABLE IF NOT EXISTS public.waypoint
+CREATE TABLE IF NOT EXISTS magic_managed.waypoint
 (
   pk               INTEGER                  GENERATED ALWAYS AS IDENTITY
     CONSTRAINT waypoint_pk PRIMARY KEY,
@@ -136,23 +142,23 @@ CREATE TABLE IF NOT EXISTS public.waypoint
 );
 
 CREATE INDEX IF NOT EXISTS waypoint_geom_idx
-  ON public.waypoint USING gist (geom);
+  ON magic_managed.waypoint USING gist (geom);
 
 CREATE OR REPLACE TRIGGER waypoint_updated_at_trigger
   BEFORE INSERT OR UPDATE
-  ON waypoint
+  ON magic_managed.waypoint
   FOR EACH ROW
   EXECUTE FUNCTION set_updated_at();
 
 CREATE OR REPLACE TRIGGER waypoint_updated_by_trigger
   BEFORE INSERT OR UPDATE
-  ON waypoint
+  ON magic_managed.waypoint
   FOR EACH ROW
   EXECUTE FUNCTION set_updated_by();
 
 -- ROUTE CONTAINER
 
-CREATE TABLE IF NOT EXISTS public.route_container
+CREATE TABLE IF NOT EXISTS magic_managed.route_container
 (
   pk         INTEGER                  GENERATED ALWAYS AS IDENTITY
     CONSTRAINT route_pk PRIMARY KEY,
@@ -164,26 +170,26 @@ CREATE TABLE IF NOT EXISTS public.route_container
 
 CREATE OR REPLACE TRIGGER route_container_updated_at_trigger
   BEFORE INSERT OR UPDATE
-  ON route_container
+  ON magic_managed.route_container
   FOR EACH ROW
   EXECUTE FUNCTION set_updated_at();
 
 CREATE OR REPLACE TRIGGER route_container_updated_by_trigger
   BEFORE INSERT OR UPDATE
-  ON route_container
+  ON magic_managed.route_container
   FOR EACH ROW
   EXECUTE FUNCTION set_updated_by();
 
 -- ROUTE WAYPOINT
 
-CREATE TABLE IF NOT EXISTS public.route_waypoint
+CREATE TABLE IF NOT EXISTS magic_managed.route_waypoint
 (
     pk           INTEGER                  GENERATED ALWAYS AS IDENTITY
       CONSTRAINT route_waypoint_pk PRIMARY KEY,
     route_pid    UUID                     NOT NULL
-      CONSTRAINT route_waypoint_route_pid_fk REFERENCES route_container(pid) ON DELETE CASCADE,
+      CONSTRAINT route_waypoint_route_pid_fk REFERENCES magic_managed.route_container(pid) ON DELETE CASCADE,
     waypoint_pid UUID                     NOT NULL
-      CONSTRAINT route_waypoint_waypoint_pid_fk REFERENCES waypoint(pid) ON DELETE CASCADE,
+      CONSTRAINT route_waypoint_waypoint_pid_fk REFERENCES magic_managed.waypoint(pid) ON DELETE CASCADE,
     sequence     INTEGER                  NOT NULL,
     updated_at   TIMESTAMP WITH TIME ZONE NOT NULL DEFAULT now(),
     updated_by   TEXT                     NOT NULL DEFAULT 'unknown'
@@ -191,59 +197,58 @@ CREATE TABLE IF NOT EXISTS public.route_waypoint
 
 CREATE OR REPLACE TRIGGER route_waypoint_updated_at_trigger
   BEFORE INSERT OR UPDATE
-  ON route_waypoint
+  ON magic_managed.route_waypoint
   FOR EACH ROW
   EXECUTE FUNCTION set_updated_at();
 
 CREATE OR REPLACE TRIGGER route_waypoint_updated_by_trigger
   BEFORE INSERT OR UPDATE
-  ON route_waypoint
+  ON magic_managed.route_waypoint
   FOR EACH ROW
   EXECUTE FUNCTION set_updated_by();
 
 -- ROUTE
 
-CREATE OR REPLACE VIEW route AS
+CREATE OR REPLACE VIEW magic_managed.route AS
     WITH route_geom AS (
         SELECT rw.route_pid, st_makeline(w.geom ORDER BY rw.sequence)::geometry(LINESTRING, 4326) AS geom
-        FROM route_waypoint AS rw
-        JOIN waypoint w ON w.pid = rw.waypoint_pid
+        FROM magic_managed.route_waypoint AS rw
+        JOIN magic_managed.waypoint w ON w.pid = rw.waypoint_pid
         GROUP BY rw.route_pid
     )
     SELECT rc.pid, rc.id, rg.geom
-    FROM route_container AS rc
+    FROM magic_managed.route_container AS rc
     FULL OUTER JOIN route_geom rg ON rc.pid = rg.route_pid;
 
 CREATE OR REPLACE FUNCTION route_insert() RETURNS TRIGGER AS $$
 BEGIN
-    INSERT INTO route_container (id) VALUES (NEW.id);
-    INSERT INTO route_waypoint (route_pid, waypoint_pid, sequence)
+    INSERT INTO magic_managed.route_container (id) VALUES (NEW.id);
+    INSERT INTO magic_managed.route_waypoint (route_pid, waypoint_pid, sequence)
         (
             SELECT rc.pid as route_pid, w.pid as waypoint_pid, points.path[1] AS sequence
             FROM ST_dumppoints(NEW.geom) AS points
-            JOIN waypoint w ON ST_DWithin(ST_Transform(w.geom, 3857) , ST_Transform(points.geom, 3857), 1000)
-            JOIN route_container rc ON rc.id = NEW.id
+            JOIN magic_managed.waypoint w ON ST_DWithin(ST_Transform(w.geom, 3857) , ST_Transform(points.geom, 3857), 1000)
+            JOIN magic_managed.route_container rc ON rc.id = NEW.id
         );
     RETURN NEW;
 END;
 $$ LANGUAGE plpgsql;
 
 CREATE TRIGGER route_insert_trigger
-INSTEAD OF INSERT ON route
+INSTEAD OF INSERT ON magic_managed.route
 FOR EACH ROW EXECUTE FUNCTION route_insert();
-
 
 CREATE OR REPLACE FUNCTION route_update() RETURNS TRIGGER AS $$
 BEGIN
-    UPDATE route_container SET id = NEW.id
+    UPDATE magic_managed.route_container SET id = NEW.id
         WHERE pid = OLD.pid;
     DELETE FROM route_waypoint
         WHERE route_pid = OLD.pid AND NEW.geom IS NOT NULL;
-    INSERT INTO route_waypoint (route_pid, waypoint_pid, sequence)
+    INSERT INTO magic_managed.route_waypoint (route_pid, waypoint_pid, sequence)
         (
             SELECT OLD.pid as route_pid, w.pid as waypoint_pid, points.path[1] AS sequence
             FROM ST_dumppoints(NEW.geom) AS points
-            JOIN waypoint w ON w.geom = points.geom
+            JOIN magic_managed.waypoint w ON w.geom = points.geom
             WHERE NEW.geom IS NOT NULL
         );
     RETURN NEW;
@@ -251,18 +256,17 @@ END;
 $$ LANGUAGE plpgsql;
 
 CREATE TRIGGER route_update_trigger
-INSTEAD OF UPDATE ON route
+INSTEAD OF UPDATE ON magic_managed.route
 FOR EACH ROW EXECUTE FUNCTION route_update();
-
 
 CREATE OR REPLACE FUNCTION route_delete() RETURNS TRIGGER AS $$
 BEGIN
-    DELETE FROM route_waypoint WHERE route_pid = OLD.pid;
-    DELETE FROM route_container WHERE pid = OLD.pid;
+    DELETE FROM magic_managed.route_waypoint WHERE route_pid = OLD.pid;
+    DELETE FROM magic_managed.route_container WHERE pid = OLD.pid;
     RETURN OLD;
 END;
 $$ LANGUAGE plpgsql;
 
 CREATE TRIGGER route_delete_trigger
-INSTEAD OF DELETE ON route
+INSTEAD OF DELETE ON magic_managed.route
 FOR EACH ROW EXECUTE FUNCTION route_delete();
